* add missing queries
