+++ zgw/openklant/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,2 @@
 nl.nlportal.openklant.autoconfigure.OpenKlantAutoConfiguration
+nl.nlportal.openklant.autoconfigure.OpenKlantGraphqlAutoConfiguration
