+++ zgw/openklant/src/test/resources/config/graphql/createUserPartij.gql
@@ -1,0 +1,16 @@
+mutation {
+    createUserPartij(partijRequest: { type: PERSOON, indicatieActief: true, indicatieGeheimhouding: true, persoonsIdentificatie: {contactnaam: { voornaam: "Bob", voorvoegselAchternaam: "de", achternaam: "Bouwer" }, volledigeNaam: "Bob de Bouwer" } }) {
+        type
+        indicatieActief
+        indicatieGeheimhouding
+        persoonsIdentificatie {
+            contactnaam {
+                voorletters
+                voornaam
+                voorvoegselAchternaam
+                achternaam
+            }
+            volledigeNaam
+        }
+    }
+}
