+++ zgw/openklant/src/test/resources/config/graphql/openKlant2TypeIntrospection.gql
@@ -1,0 +1,7 @@
+query {
+    __type(name: "PartijResponse") {
+        kind
+        name
+        description
+    }
+}
